+++ app/(redux)/redux-demo/add-post/page.tsx
@@ -1,4 +1,4 @@
-import { AddPostForm } from '@/components/atomic-design/organisms/AddPostForm/AddPostForm';
+import { AddPostForm } from '@/components/atomic-design/organisms/AddPostForm';
 
 export const Page = () => {
   return <AddPostForm />;
