+++ app/(redux)/redux-demo/edit-post/[id]/page.tsx
@@ -1,4 +1,4 @@
-import EditPostForm from '@/components/atomic-design/organisms/EditPostForm/EditPostForm';
+import EditPostForm from '@/components/atomic-design/organisms/EditPostForm';
 
 export default async function Page({ params }: { params: { id: string } }) {
   const postId = params.id;
