+++ app/(redux)/redux-demo/post/[id]/page.tsx
@@ -1,4 +1,4 @@
-import SinglePost from '@/components/atomic-design/organisms/SinglePost/SinglePost';
+import SinglePost from '@/components/atomic-design/organisms/SinglePost';
 
 export default async function Page({ params }: { params: { id: string } }) {
   const postId = params.id;
