+++ components/atomic-design/organisms/AddPostForm/index.tsx
@@ -1,9 +1,9 @@
 'use client';
 
 import React, { useState } from 'react';
-import Subtitle from '../../atoms/Subtitle/Subtitle';
+import Subtitle from '../../atoms/Subtitle';
 import { postAdded } from 'store/state-management/redux/posts/postsSlice';
-import PostForm from '../PostForm/PostForm';
+import PostForm from '../PostForm';
 import { useAppDispatch } from 'store/state-management/redux/hook';
 
 export const AddPostForm = () => {
