+++ components/atomic-design/organisms/PostCard/index.styles.ts
