+++ components/atomic-design/organisms/PostForm/index.styles.ts
