+++ components/atomic-design/organisms/PostsList/index.tsx
@@ -2,8 +2,8 @@
 
 import React from 'react';
 import { selectAllPosts } from 'store/state-management/redux/posts/postsSlice';
-import { PostCard } from '../PostCard/PostCard';
-import Subtitle from '../../atoms/Subtitle/Subtitle';
+import { PostCard } from '../PostCard';
+import Subtitle from '../../atoms/Subtitle';
 import { useAppSelector } from 'store/state-management/redux/hook';
 
 export const PostsList = () => {
@@ -22,3 +22,5 @@ export const PostsList = () => {
     </section>
   );
 };
+
+export default PostsList;
