+++ components/atomic-design/organisms/SinglePost/index.tsx
@@ -4,7 +4,7 @@ import { notFound } from 'next/navigation';
 import { useSelector } from 'react-redux';
 import { selectPostById } from 'store/state-management/redux/posts/postsSlice';
 import { RootState } from 'store/state-management/redux/store';
-import { PostCard } from '../PostCard/PostCard';
+import { PostCard } from '../PostCard';
 import { Button } from '../../atoms/Button-SC';
 import Link from 'next/link';
 
@@ -17,7 +17,7 @@ export const SinglePost = ({ id }: { id: string }) => {
 
   return (
     <>
-      <PostCard title={post.title} content={post.content} />
+      <PostCard title={post.title} content={post.content} id={post.id} />
       <Link href={`/redux-demo/edit-post/${id}`}>
         <Button label="edit post" size="small" style={{ marginTop: '10px' }} />
       </Link>
